pyproject.toml
[project]
name = "dbspre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "psycopg2-binary>=2.9.11",
    "pyproj>=3.7.2",
    "python-dotenv>=1.2.2",
    "shapely>=2.1.2",
    "streamlit>=1.56.0",
    "traci>=1.26.0",
]